+++ components/stream-subjects/retrieve.php
@@ -25,16 +25,19 @@
 		<div class="panel panel-primary">
 			<div class="panel-heading">
 				<div class="panel-title">
-					<h4 class="panel-title"><?php echo $streamName?> Subjects</h4>
+					<h4 class="panel-title"><?=$streamName?> Subjects</h4>
 				</div>
 			</div>
 			<div class="panel-body">
-				<a class="btn btn-primary" data-keyboard="false" href="/bookshop?action=new-stream-subject&id=<?php echo $streamId; ?>&name=<?php echo $streamName; ?>"><span class="glyphicon glyphicon-plus"></span>&nbsp;Add New</a><br /><br />
-
+				<?php if ($_SESSION['loggedRole'] == 'admin'): ?>
+					<a class="btn btn-primary" data-keyboard="false" href="/bookshop?action=new-stream-subject&id=<?php echo $streamId; ?>&name=<?php echo $streamName; ?>"><span class="glyphicon glyphicon-plus"></span>&nbsp;Add New</a><br /><br />
+				<?php endif; ?>
 				<table id="subjectsTable" class="table table-bordered table-hover">
 					<thead>
 						<tr>
-							<th></th>
+							<?php if ($_SESSION['loggedRole'] == 'admin'): ?>
+								<th></th>
+							<?php endif; ?>
 							<th>Name</th>
 							<th>Description</th>
 						</tr>
@@ -43,9 +46,11 @@
 						
 						<?php foreach($subjects as $row): ?>
 							<tr>
-								<td>
-									<a style="color: #FF0000;" href="/bookshop?action=delete-stream-subject&id=<?php echo $row['id']; ?>"><span class='glyphicon glyphicon-trash' aria-hidden='true'></span>Remove</a>&nbsp;&nbsp;
-								</td>
+								<?php if ($_SESSION['loggedRole'] == 'admin'): ?>
+									<td>
+										<a style="color: #FF0000;" href="/bookshop?action=delete-stream-subject&id=<?php echo $row['id']; ?>"><span class='glyphicon glyphicon-trash' aria-hidden='true'></span>Remove</a>&nbsp;&nbsp;
+									</td>
+								<?php endif; ?>
 								<td><?=$row['name']?></td>
 								<td><?=$row['description']?></td>
 							</tr>
