+++ components/streams/retrieve.php
@@ -13,12 +13,15 @@
 				</div>
 			</div>
 			<div class="panel-body">
-				<a class="btn btn-primary" data-keyboard="false" href="/bookshop?action=new-stream"><span class="glyphicon glyphicon-plus"></span>&nbsp;Add New</a><br /><br />
-
+				<?php if ($_SESSION['loggedRole'] == 'admin'): ?>
+					<a class="btn btn-primary" data-keyboard="false" href="/bookshop?action=new-stream"><span class="glyphicon glyphicon-plus"></span>&nbsp;Add New</a><br /><br />
+				<?php endif; ?>
 				<table id="streamsTable" class="table table-bordered table-hover">
 					<thead>
 						<tr>
-							<th></th>
+							<?php if ($_SESSION['loggedRole'] == 'admin'): ?>
+								<th></th>
+							<?php endif; ?>
 							<th>Name</th>
 							<th>Description</th>
 						</tr>
@@ -34,11 +37,13 @@
 
 						<?php foreach($streams as $row): ?>
 							<tr>
-								<td>
-									<a href="/bookshop?action=edit-stream&id=<?php echo $row['id']; ?>"><span class='glyphicon glyphicon-edit' aria-hidden='true'></span>Edit</a>&nbsp;&nbsp;
-									<a style="color: #FF0000;" href="/bookshop?action=delete-stream&id=<?php echo $row['id']; ?>"><span class='glyphicon glyphicon-trash' aria-hidden='true'></span>Remove</a>&nbsp;&nbsp;
-									<a href="/bookshop?action=stream-subjects&id=<?php echo $row['id']; ?>&name=<?php echo $row['name']; ?>"><span class='glyphicon glyphicon-link' aria-hidden='true'></span>Subjects</a>
-								</td>
+								<?php if ($_SESSION['loggedRole'] == 'admin'): ?>
+									<td>
+										<a href="/bookshop?action=edit-stream&id=<?php echo $row['id']; ?>"><span class='glyphicon glyphicon-edit' aria-hidden='true'></span>Edit</a>&nbsp;&nbsp;
+										<a style="color: #FF0000;" href="/bookshop?action=delete-stream&id=<?php echo $row['id']; ?>"><span class='glyphicon glyphicon-trash' aria-hidden='true'></span>Remove</a>&nbsp;&nbsp;
+										<a href="/bookshop?action=stream-subjects&id=<?php echo $row['id']; ?>&name=<?php echo $row['name']; ?>"><span class='glyphicon glyphicon-link' aria-hidden='true'></span>Subjects</a>
+									</td>
+								<?php endif; ?>
 								<td><?=$row['name']?></td>
 								<td><?=$row['description']?></td>
 							</tr>
